chore: refactor product controller tests
